requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langchain-groq>=0.2.0
faiss-cpu>=1.12.0
pymupdf>=1.24.0
rank_bm25>=0.2.2
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
tiktoken>=0.7.0
pydantic>=2.7.0
sentence-transformers